+++ components/index.ts
@@ -1,10 +1,7 @@
-export * from './Author/Author';
 export * from './Blog/Blog';
-export * from './ArticleImg/ArticleImg';
 export * from './SearchInput/SearchInput';
 export * from './Dropdown/Dropdown';
 export * from './ApiError/ApiError';
-export * from './ApiErrorBoundary/ApiErrorBoundary';
 export * from './ImageModal';
 export * from './Follower/Follower';
 export * from './PasswordValidityLevel/PasswordValidityLevel';
