+++ components/molecules/ArticleImg/ArticleImg.types.ts
