+++ components/molecules/index.ts
@@ -13,3 +13,4 @@ export * from './no-article/NoArticle';
 export * from './auth-button/AuthButton';
 export * from './labels-selector/LabelsSelector';
 export * from './upload-status/UploadStatus';
+export * from './ApiErrorBoundary/ApiErrorBoundary';
