+++ frontends/user-articles/published/PublishedArticles.tsx
@@ -1,6 +1,5 @@
-import { ApiErrorBoundary } from 'components';
 import { Button, Spinner, StorysetImage } from 'components/lib';
-import { Article, LoadMoreButton } from 'components/molecules';
+import { ApiErrorBoundary, Article, LoadMoreButton } from 'components/molecules';
 import { useAppRouter } from 'hooks';
 import { FC, useCallback } from 'react';
 import { useBlogArticles, useCreateArticle } from 'store/clients/article';
